requirements.txt
asn1crypto==0.24.0
boto3==1.9.82
botocore==1.12.82
cachetools==3.1.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
cloudstorage==0.9.0
cryptography==2.5
docutils==0.14
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
google-api-core==1.7.0
google-auth==1.6.3
google-cloud-core==0.29.1
google-cloud-pubsub==0.39.1
google-cloud-storage==1.14.0
google-resumable-media==0.3.2
googleapis-common-protos==1.5.8
grpc-google-iam-v1==0.11.4
grpcio==1.18.0
gunicorn==19.9.0
idna==2.8
inflection==0.3.1
itsdangerous==1.1.0
Jinja2==2.10
jmespath==0.9.3
MarkupSafe==1.1.0
protobuf==3.6.1
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycparser==2.19
PyMySQL==0.9.3
pyOpenSSL==19.0.0
PySocks==1.6.8
python-dateutil==2.7.5
python-magic==0.4.15
pytz==2018.9
requests==2.21.0
rsa==4.0
s3transfer==0.1.13
six==1.12.0
SQLAlchemy==1.2.18
urllib3==1.24.1
webapp2==2.5.2
Werkzeug==0.14.1
win-inet-pton==1.0.1
wincertstore==0.2
